RealTimeVideoFrameClassificationFramework/requirements.txt
Pillow
PyOpenGL
pygame
pandas
seaborn
scikit-learn
joblib
opencv-python
pyyaml
onnxruntime
vidgear
numba
keras_cv
tensorflow
tensorflow-addons
tensorflow-probability